requirements.txt
pyyaml==5.3.1
cx-Oracle=8.0.1
psycopg2=2.8.6
